Add winner function as display message
